backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.0
numpy==1.26.4
lightgbm==4.1.0
scikit-learn==1.5.0
requests==2.31.0
